[css-flexbox] Update our cached heights when laying out our child
If we lay out our child before calling computeInnerFlexBaseForChild, we did
not update the cached sizes in m_intrinsicSizeAlongMainAxis. This is very bad
because that means we will only update that cache the first time we call
computeInnerFlexBaseForChild (when the size is not in the hashmap yet),
but we will never update it because needsLayout() will always be false
in that function. So we have to make sure to update the cache earlier.

R=leviw@chromium.org
BUG=550293

Review URL: https://codereview.chromium.org/1419993008

Cr-Original-Commit-Position: refs/heads/master@{#357686}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: f19b5e12221e4d3ebd59ef7d0a37f5465721c621