Add runtime enabled feature for Experimental Framework
The EF will likely be used with APIs that start out
behind the experimental-web-platform-features flag.
Until the framework is further along, it should be
possible to enable/disable the framework independently
from APIs that are using it.

In subsequent CL(s), it will be possible for an API to
integrate with the framework, but still be functional
even if the framework is disabled.

For now, the runtime enabled feature is intentionally off by default.

BUG=543141

Review URL: https://codereview.chromium.org/1429103003

Cr-Original-Commit-Position: refs/heads/master@{#357966}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: d2932f32226717085536a58ab0e0fbd8509eadb0