Let RenderFrameImpl set navigationStart based on CommonNavigationParams
This entails setting the timestamp before we set it in blink, so DocumentLoadTiming had to be modified to allow for that. We set the value in didCreateDataSource.

didCreateDataSource was also refactored slightly so that same-page navigations don't need to call it anymore.

This is 2/3 of the effort to add a navigation_start timestamp to DidStartProvisional load. The full CL we're breaking up is here:
https://codereview.chromium.org/1425823002/

BUG=548335

Review URL: https://codereview.chromium.org/1423453007

Cr-Original-Commit-Position: refs/heads/master@{#358144}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 5f61ecb107fd7118621a99fe03b094f37ce04394