bluetooth: Use GattServicesDiscovered function
The GattServicesDiscovered function is called when all services
have been discovered for a device. We can use it to remove
the timeout when connecting to a device. Meaning instead of waiting
N seconds before reading the services we check if the service is present
if not we wait until all services have been discovered and then
check if the service is present again.

BUG=491399

Review URL: https://codereview.chromium.org/1419623006

Cr-Original-Commit-Position: refs/heads/master@{#358175}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 31111f33ab1914334ded0fc0560905fa63de8e04