Disable eager tracing for thread-terminated GCs.
The Oilpan marking phase enables and sets up a stack limit
that the eager tracing calls should not exceed while visiting the live
object graph. The marking phase is single threaded. The stack limit
is reset and eager marking disabled once the marking phase has
completed -- thread local weak processing (and its slight amount of
tracing) is run with it disabled.

Do not enable such an eager marking stack limit for the GCs that are run
for a thread's heap when the thread is terminated, as this risks conflicting
with any thread-local weak processing that is still underway. The slight
performance gain that eager marking brings is not an issue for such GCs.

R=haraken
BUG=420515

Review URL: https://codereview.chromium.org/1429273002

Cr-Original-Commit-Position: refs/heads/master@{#358273}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 76c2d557fcab24b4d76659301006d0d8f5e196e8