Cache OriginAccessEntry in Document
Document::firstPartyForCookies is called for all resource requests,
and an OriginAccessEntry is instantiated for each invocation.
That costs around 60us, 10% of the time for each issuing resource request.

This CL adds a cache for it to save that time.

BUG=540988

Review URL: https://codereview.chromium.org/1405213005

Cr-Original-Commit-Position: refs/heads/master@{#358280}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 1f3a3add3252cb14e53e21ee656923649b927cfb