Correct sampling of timeline in setValueCurveAtTime
The interpolation of the curve in setValueCurveAtTime was slightly
off.  This was partially due to the implementation and also due to an
error in the spec.  The spec is being updated with the correct
formula: https://github.com/WebAudio/web-audio-api/pull/658

BUG=441471
TEST=audioparam-setValueCurveAtTime.html,
     audioparam-setValueCurveAtTime-interpolation.html

Review URL: https://codereview.chromium.org/1414033011

Cr-Original-Commit-Position: refs/heads/master@{#359178}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 7f2cab7d9a683b8447074023a38f4a41fd0871f0