Fix a crash when revalidation succeeded for a raw resource without body.
When revalidation is succeeded for a resource, the body contents is dispatched
to its clients. The existing code wrongly assumes that |m_data| is non-null,
but it could be null if the original resource doesn't have body.

BUG=549022

Review URL: https://codereview.chromium.org/1432613005

Cr-Original-Commit-Position: refs/heads/master@{#359455}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 3e0346911c19478b179baee0cf22c7a57521aa52