[webnfc] NFC is not required to be ActiveDOMObject.
NFC object is a member of NavigatorNFC that is supplement for Navigator.
Therefore, NFC object will be alive as long as Navigator object exists.
Because of that, inheritance from ActiveDOMObject is not required.

PageLifeCycleObserer is used to fullfill visibility use-case.
https://w3c.github.io/web-nfc/#handling-window-visibility-and-focus

BUG=520391

Review URL: https://codereview.chromium.org/1439063002

Cr-Original-Commit-Position: refs/heads/master@{#359531}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: e2cea06427fc1904111f7a0dc28d1f052d9f52a5