Remove the IPCEcho service and WebDOMCustomEvent.
IPCEcho was to experiment with the performance of chrome ipc vs mojo ipc
a long time ago, but there's no perf tests that use it, so the code just
rots behind the flag. It's also the only user of WebDOMCustomEvent which
we can now remove too.

Lets just delete this code, in the future if we want a way to test
the performance of mojo IPCs we can add it back with automated tests.

This patch also removes WebDOMCustomEvent.

Review URL: https://codereview.chromium.org/1438403004

Cr-Original-Commit-Position: refs/heads/master@{#359568}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 4b0602d9f07efe1e616d048747268dd4c20e15f9