Remove clientWidth call during MediaControls::reset.
Avoid calling clientWidth early to get the control panel width.
It can cause a crash when the HitTestCache is cleared, if the
cache was the last holder to the media element.  Normally, this
can't happen, but a callback from the WebMediaPlayer can be
scheduled between the cache becoming the last reference and the
next layout.

BUG=531259

Review URL: https://codereview.chromium.org/1413103005

Cr-Original-Commit-Position: refs/heads/master@{#359735}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: f003b7d5f0f8667873a8f152d5a06512d3bded9c