TestImage is ref-counted: it should have public create() methods
Class seems to be a copy of the ImageLayerChromiumTest.cpp class of
the same name, but oddly has public constructors and is ref-counted
which is not really per style. Privatise constructors, add a helper
to create the SkSuface.

BUG=439655

Review URL: https://codereview.chromium.org/1446133002

Cr-Original-Commit-Position: refs/heads/master@{#359943}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 81c5309cfd874d6d8a4f1ef8134f80c89d8e4c98