WebAudio: Allow non-audio threads to call DefferedTaskHandler::lock
Currently DefferedTaskHandler::lock has ASSERT(isMainThread()), but this is not correct
because a GCing thread (which may not be the main thread) calls DefferedTaskHandler::lock
when tracing an AbstractAudioContext.

What we really want to assert in DefferedTaskHandler::lock is that it is not called
in a real-time audio thread. Thus this CL changes the assert to ASSERT(!isAudioThread()).

BUG=556478

Review URL: https://codereview.chromium.org/1445393002

Cr-Original-Commit-Position: refs/heads/master@{#360016}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 91f4d0ce2173e469fcd4c9f77c639a68297da317