Enable MinorGCUnmodifiedWrapperVisitor in workers
Currently MinorGCUnmodifiedWrapperVisitor is disabled in workers.
Consequently, hasPendingActivity is ignored and a bunch of wrappers are wrongly collected
when a V8 minor GC is triggered in workers. This has caused a bunch of undeterministic crashes
in workers.

BUG=550632,553769

Review URL: https://codereview.chromium.org/1450293002

Cr-Original-Commit-Position: refs/heads/master@{#360032}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 84814a37400d1cfa6707e7bb06dddeacd492d7f2