Support 'URLSearchParams' as the body of a Request
As specified in Fetch, this patch allows 'URLSearchParams' to be passed
into the Request constructor, or into the RequestInit parameter of a
call to 'fetch()'. The object will be serialized as a urlencoded string,
and used as the body of the resulting request.

BUG=303152
NOTRY=true

Review URL: https://codereview.chromium.org/1456553002

Cr-Original-Commit-Position: refs/heads/master@{#360326}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 03eaf889ebe699c3a5af069d4a2ef2206c6431ad