[css-grid] Store lines instead of tracks in GridResolvedPosition
Due to the new feature that allows to create implicit tracks before the
explicit ones, we will need to use lines instead of tracks in the
code to be able to implement it properly.

This is just a first simple patch using lines instead of tracks in
GridResolvedPosition. It modifies the code that was using it, as it was
considering that the resolvedFinalPosition was a track index and
not a line index.

So if we've an item positioned like:
    grid-column: 2 / 5;
    grid-row: 1 / span 2;

Before we were storing this information on the GridSpan:
    * columns:
        * resolvedInitialPosition: 1
        * resolvedFinalPosition:  3
    * rows:
        * resolvedInitialPosition: 0
        * resolvedFinalPosition:  1

And now we're storing:
    * columns:
        * resolvedInitialPosition: 1
        * resolvedFinalPosition:  4
    * rows:
        * resolvedInitialPosition: 0
        * resolvedFinalPosition:  2

No new tests, no change of behavior.

BUG=444011

Review URL: https://codereview.chromium.org/1451883002

Cr-Original-Commit-Position: refs/heads/master@{#360361}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 2031b1892af1ca54c5b85660f6f369979e548fdb