[PartitionAlloc] Fix type names in fast malloc macro
The type name in the |USING_FAST_MALLOC| macro is currently not used,
which allowed a few typos to sneak in. This fixes those to prevent
compile errors in case the type name is used in the future.

Review URL: https://codereview.chromium.org/1461743004

Cr-Original-Commit-Position: refs/heads/master@{#360569}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: ddc763cdec5788f27da528d35ab31a3428a1fc84