Update tests for Android results
Update the thresholds so that the layout tests pass on Android on arm.

audioparam-setTargetAtTime needs an updated threshold to pass on arm
and arm64.

audioparam-setValueCurveAtTime-interpolation doesn't produce the
expected 0 on arm64, but is very close, so update the test to use a
threshold instead of an exact value. (See issue 558563).

audio-testing updated so Should.beCloseTo will use an absolute
threshold if the target value is 0.

Also fixes issue 557444 since this was how it was discovered.

BUG=557433, 557444
TEST=audioparam-setTargetAtTime, audioparam-setValueCurveAtTime-interpolation

Review URL: https://codereview.chromium.org/1457153003

Cr-Original-Commit-Position: refs/heads/master@{#360694}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: cecefefed0cfa7744c05c0217655bbcd3bb15222