We mustn't finalize sweeping when V8 started collecting garbage.
The finalizers might call into V8 which is not safe to do. There's
actually a script forbidden scope on the stack in the prologue, but it's
created right after the call to finialize sweeping, so it didn't catch
this before.

BUG=none
R=sigbjornf@opera.com,haraken@chromium.org

Review URL: https://codereview.chromium.org/1467493002

Cr-Original-Commit-Position: refs/heads/master@{#360834}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 93276fb7983b7794cc533581462f7c39e623dea5