Add context information to BidiResolver.isolatedRuns()
This patch is a port of http://wkb.ug/149153, which is a port of a fix
for crbug.com/520282.

The WebKit patch has additional changes to carry more information of
isolated runs along with BidiRun. As we find we need to carry more
context with isolated runs in crbug.com/391260, this change is useful.

BUG=520282

Review URL: https://codereview.chromium.org/1468103002

Cr-Original-Commit-Position: refs/heads/master@{#361258}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: f5c7aca8e2c3f4ae965f3c6098e4c6eb00366b30