[DevTools]Check the existence of WebViewClient in StorageClientImpl::createSessionStorageNamespace()
WebViewImpl.m_client is null in SharedWorker and ServiceWorker. But
createSessionStorageNamespace() doesn't check the existence of it. So when the
user clicks a domain in "Session Storage" in DevTools's Resources tab, the
renderer will crash.

This cl changes it to check of it and return nullptr.

BUG=421384

Review URL: https://codereview.chromium.org/1469293002

Cr-Original-Commit-Position: refs/heads/master@{#361559}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 50ae488d668dfce2942b868dab6167f7b484399d